Generates a cluster definition

Defines a Beowulf cluster for makeCluster from the IPs of the machines in the cluster, the number of cores of each machine, and the user name.

cluster_specification(
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]]
)

Arguments

cluster.ips
Character vector with the IPs of the machines in the cluster. The first machine will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed. Default: NULL.

cluster.cores
Numeric integer vector, number of cores on each machine. Default: NULL.

cluster.user
Character string, name of the user (should be the same throughout machines), Defaults to the current system user.

Value

A list ready to be used as input for the spec argument of the function makeCluster.

Details

This function is used internally by several other functions in the package, but can be useful for a user working with foreach to parallelize loops in a small cluster.

Examples

 cluster.spec <- cluster_specification(
   cluster.ips = c("10.0.1.40", "10.0.1.41"),
   cluster.cores = c(6, 4),
   cluster.user = Sys.info()[["user"]]
 )
 cluster.spec

#> [[1]]
#> [[1]]$host
#> [1] "10.0.1.40"
#> 
#> [[1]]$user
#> [1] "blas"
#> 
#> 
#> [[2]]
#> [[2]]$host
#> [1] "10.0.1.40"
#> 
#> [[2]]$user
#> [1] "blas"
#> 
#> 
#> [[3]]
#> [[3]]$host
#> [1] "10.0.1.40"
#> 
#> [[3]]$user
#> [1] "blas"
#> 
#> 
#> [[4]]
#> [[4]]$host
#> [1] "10.0.1.40"
#> 
#> [[4]]$user
#> [1] "blas"
#> 
#> 
#> [[5]]
#> [[5]]$host
#> [1] "10.0.1.40"
#> 
#> [[5]]$user
#> [1] "blas"
#> 
#> 
#> [[6]]
#> [[6]]$host
#> [1] "10.0.1.40"
#> 
#> [[6]]$user
#> [1] "blas"
#> 
#> 
#> [[7]]
#> [[7]]$host
#> [1] "10.0.1.41"
#> 
#> [[7]]$user
#> [1] "blas"
#> 
#> 
#> [[8]]
#> [[8]]$host
#> [1] "10.0.1.41"
#> 
#> [[8]]$user
#> [1] "blas"
#> 
#> 
#> [[9]]
#> [[9]]$host
#> [1] "10.0.1.41"
#> 
#> [[9]]$user
#> [1] "blas"
#> 
#> 
#> [[10]]
#> [[10]]$host
#> [1] "10.0.1.41"
#> 
#> [[10]]$user
#> [1] "blas"
#> 
#>