Gets model residuals

Returns the residuals of models fitted with rf(), rf_repeat(), or rf_spatial().

Usage

get_residuals(model)

Arguments

model

A model fitted with rf(), rf_repeat(), or rf_spatial().

Value

A vector with model residuals, or the median of model residuals across repetitions if the model was fitted with rf_repeat().

Examples

#load example data
data(plant_richness_df)

#fit random forest model
rf.model <- rf(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  n.cores = 1,
  verbose = FALSE
)

#getting vector with residuals
x <- get_residuals(rf.model)
x
#>   [1]  -254.968633  -193.200133  -136.609767   820.822033  -523.103833
#>   [6]  -184.907367   229.445167  1280.537000  -163.265433   -60.073267
#>  [11]  -298.493467  -366.725633  -227.701633    19.962833  2385.109600
#>  [16]   838.289800   232.174800  -242.180000 -1311.985433  -596.654400
#>  [21]  1423.255567   852.787700  3221.384667  -136.068700   -90.540200
#>  [26]   842.596433  -358.489400   -63.694900  -335.051833  -444.897733
#>  [31] -1320.048267  1405.227167   459.439667   149.274267 -1751.614700
#>  [36]  -409.953033  -440.141033  -235.778200   -90.153333  -178.831133
#>  [41]  -293.039667   454.447500  -866.747533  -369.226867  -475.163167
#>  [46]  -898.645200  -472.216433  -431.208700  -369.125700   602.568900
#>  [51]  -861.706533  -207.787433   -88.618733  -196.610667  -211.962333
#>  [56]   900.257300  -371.229100  -820.134200  -799.495233  -121.653133
#>  [61]  -410.183800  -875.763433   805.400433   617.456300   -28.727467
#>  [66]    -8.869500  1440.725200   448.554000  -283.470967   -60.459700
#>  [71]   259.958467     9.104300  2500.725933  -120.138667  -724.273033
#>  [76]  -700.492600   -79.190367  -258.087767  -960.763567  -533.337400
#>  [81]   -28.864600  -235.020533  -668.191533  -447.755067   -38.072500
#>  [86]   -73.036867  -350.080700  -469.194800    23.674267  -335.998333
#>  [91]  -470.906200  -460.636867  -280.563633  1243.453100  -192.150600
#>  [96]  -132.941667  1616.247000   510.448367  1186.750900  -141.265100
#> [101]  -117.699500    26.381967    51.658733 -1080.913900   252.572500
#> [106]  -543.487967   344.684767  -727.006967   220.706133  -529.318767
#> [111]  -645.124433   216.102833  -559.563333 -1708.861800   536.209600
#> [116]  -187.573733  -363.276067  -545.216533  -363.570400    10.660500
#> [121]  -884.240767  1226.217833    28.118567   136.374467  -863.170967
#> [126]  -597.112167  -513.559367  -268.965700  -157.453267  -544.782167
#> [131]  -519.709900  -460.508933   -63.730633  -352.361067   954.792967
#> [136]  1818.107000   -24.502167  -243.695100  -545.870467   -69.256067
#> [141]   832.955800  -170.535767    -7.408067  -195.002567   277.600300
#> [146]  -855.318300  -314.143700 -1230.285100  -149.879333  -319.389233
#> [151]  -433.372167  -538.959900  -148.137533   -10.172233   272.193833
#> [156]   958.315767  -186.035633  -412.926867  -948.300833  1451.952867
#> [161]  -749.714300   134.570967   267.234733  -607.510333  8504.612300
#> [166]  -141.390433   522.607367  -125.305900  -453.326600  -380.931833
#> [171]   512.531000  -230.408267  -406.763533  -750.458500  -564.547000
#> [176]   460.725767   178.038567  2150.909700    24.013800  -268.308000
#> [181]  -903.739633  -634.417867   -96.529133     9.994933   741.042767
#> [186]  -383.845433   -56.659300    56.029067 -1585.971367   -92.294100
#> [191]   464.087200  1365.962033 -1183.391533  -408.357367   465.905067
#> [196]  -152.085100  -242.655733  4730.943567  -483.811767  -305.325333
#> [201]  -260.286567  -844.547700   487.908400  -589.960133  -584.281133
#> [206]   -40.537967  1021.612600 -1257.369333   185.195167  -358.494500
#> [211]  -366.995000   305.147667   136.382467  2718.530600  -727.679833
#> [216]  1198.293900  -678.182900  -461.116800 -1541.453033  -382.030733
#> [221]  -507.003333 -1059.172667  -571.757533    31.922067  -248.978033
#> [226]  -669.594767  -808.773200