Moran's Eigenvector Maps for different distance thresholds

Computes Moran's Eigenvector Maps of a distance matrix (using mem()) over different distance thresholds.

Usage

mem_multithreshold(
  distance.matrix = NULL,
  distance.thresholds = NULL,
  max.spatial.predictors = NULL
)

Arguments

distance.matrix

Distance matrix. Default: NULL.

distance.thresholds

Numeric vector with distance thresholds defining neighborhood in the distance matrix, Default: NULL.

max.spatial.predictors

Maximum number of spatial predictors to generate. Only useful to save memory when the distance matrix x is very large. Default: NULL.

Value

A data frame with as many rows as the distance matrix x containing positive Moran's Eigenvector Maps. The data frame columns are named "spatial_predictor_DISTANCE_COLUMN", where DISTANCE is the given distance threshold, and COLUMN is the column index of the given spatial predictor.

Details

The function takes the distance matrix x, computes its weights at difference distance thresholds, double-centers the resulting weight matrices with double_center_distance_matrix(), applies eigen to each double-centered matrix, and returns eigenvectors with positive normalized eigenvalues for different distance thresholds.

Examples

 #loading example data
 data(distance_matrix)

 #computing Moran's eigenvector maps for 0, 1000, and 2000 km
 mem.df <- mem_multithreshold(
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000)
   )
 head(mem.df)
#>   spatial_predictor_0_1 spatial_predictor_0_2 spatial_predictor_0_3
#> 1          -0.025921732           0.005220262           -0.04169686
#> 2          -0.099667898           0.053971309           -0.13244795
#> 3          -0.001047654          -0.014304635            0.04436019
#> 4          -0.016569475           0.004799116           -0.03074566
#> 5          -0.022576055           0.001959546           -0.02303677
#> 6          -0.015525218           0.002374152           -0.01979535
#>   spatial_predictor_0_4 spatial_predictor_0_5 spatial_predictor_0_6
#> 1         -0.0363323858          -0.029427120          -0.004345679
#> 2          0.3826927819           0.205013057           0.006481398
#> 3         -0.0031386301          -0.026240412           0.036601404
#> 4          0.0005170325          -0.014444892          -0.003356330
#> 5         -0.0524239401           0.006729439          -0.006797669
#> 6         -0.0338955963          -0.007884866          -0.002744701
#>   spatial_predictor_0_7 spatial_predictor_0_8 spatial_predictor_0_9
#> 1          0.0059417525          -0.003509625          -0.006296265
#> 2         -0.0338063656           0.174542054           0.024450222
#> 3         -0.0616077280           0.019577234          -0.001448598
#> 4          0.0009804302           0.010526820          -0.002889298
#> 5          0.0025571100          -0.011762268           0.004846323
#> 6          0.0029345255          -0.007054664          -0.002369704
#>   spatial_predictor_0_10 spatial_predictor_0_11 spatial_predictor_0_12
#> 1            0.002487306            0.005395791           -0.011379866
#> 2            0.005829746           -0.054961827            0.153839059
#> 3           -0.018299683            0.039246420           -0.008007024
#> 4            0.008730811            0.002146499            0.007681169
#> 5           -0.045296964           -0.039538738           -0.011040322
#> 6           -0.009715209           -0.004288805           -0.009536937
#>   spatial_predictor_0_13 spatial_predictor_0_14 spatial_predictor_0_15
#> 1            0.191604368          -0.0072850442           -0.026412798
#> 2            0.019835793          -0.0057251178           -0.062892949
#> 3           -0.014692044          -0.0201202412            0.008020865
#> 4           -0.008201179           0.0001369046            0.006005988
#> 5            0.164381811          -0.0088932051           -0.021214181
#> 6            0.178202420          -0.0060291344           -0.016907370
#>   spatial_predictor_0_16 spatial_predictor_0_17 spatial_predictor_0_18
#> 1            0.109791768             0.12559194            0.107268296
#> 2           -0.003832431             0.03668636            0.014722504
#> 3            0.048006557            -0.17744495            0.150866603
#> 4           -0.024858297            -0.01844209           -0.009470203
#> 5            0.108145269             0.02998075           -0.013711486
#> 6            0.083504053             0.06773982            0.047243449
#>   spatial_predictor_0_19 spatial_predictor_0_20 spatial_predictor_0_21
#> 1           -0.012273527            0.046229850            -0.04237342
#> 2            0.202464657            0.082229660             0.15771710
#> 3           -0.003222866            0.030550230             0.08951193
#> 4           -0.011830035            0.040433340            -0.02567659
#> 5            0.034505110           -0.178892320             0.04343963
#> 6           -0.009624029            0.008015184            -0.03834844
#>   spatial_predictor_0_22 spatial_predictor_0_23 spatial_predictor_0_24
#> 1             0.03596611             0.01118606             0.10224462
#> 2             0.04038242            -0.11437678            -0.01012395
#> 3            -0.04504866             0.13461198            -0.02047940
#> 4            -0.06842079            -0.01692236             0.10377418
#> 5            -0.13917976            -0.04071620             0.05601777
#> 6            -0.05776603            -0.02652804             0.09311909
#>   spatial_predictor_0_25 spatial_predictor_0_26 spatial_predictor_0_27
#> 1          -0.0011848037            -0.02412275          -0.0007539716
#> 2          -0.0013068838            -0.07838816           0.0235252584
#> 3          -0.0387798586            -0.07535325           0.1064394216
#> 4           0.0049076695             0.03958842           0.0382303597
#> 5          -0.0005085273            -0.04327854           0.0018750185
#> 6          -0.0021059632            -0.04019586           0.0116133257
#>   spatial_predictor_0_28 spatial_predictor_0_29 spatial_predictor_0_30
#> 1            -0.01644154           -0.052111626            0.065321865
#> 2            -0.03939033            0.003796367           -0.003727503
#> 3             0.04395511           -0.014889858           -0.004292680
#> 4            -0.01956660            0.097801181           -0.133857876
#> 5            -0.02127332           -0.066337960            0.089123320
#> 6            -0.04411492           -0.104538187            0.131117350
#>   spatial_predictor_0_31 spatial_predictor_0_32 spatial_predictor_0_33
#> 1           -0.009110300           -0.005527607           0.0024141508
#> 2            0.006294934           -0.009297452          -0.0011867862
#> 3           -0.145501124            0.048790879           0.1471245843
#> 4            0.009342614           -0.004236143           0.0002171203
#> 5           -0.008281460            0.005782912          -0.0083726169
#> 6           -0.014943405           -0.004594169          -0.0026991507
#>   spatial_predictor_0_34 spatial_predictor_0_35 spatial_predictor_0_36
#> 1           0.0995535458            -0.02277398            0.032366522
#> 2          -0.0019609284            -0.02249876            0.003705815
#> 3           0.0009205415            -0.02456633           -0.047683754
#> 4          -0.0217659867            -0.01207111           -0.002365715
#> 5          -0.0976154193             0.02262826           -0.068423826
#> 6           0.0986535382            -0.02665171            0.001086085
#>   spatial_predictor_0_37 spatial_predictor_0_38 spatial_predictor_0_39
#> 1           -0.095207273            0.001470171           0.0220490866
#> 2           -0.009986122            0.001834603           0.0072022025
#> 3           -0.030017972            0.194260820           0.0002023167
#> 4            0.003665707           -0.002490411           0.0516209027
#> 5            0.207826521            0.003515905           0.0673115372
#> 6           -0.021607620            0.005222174           0.0807527077
#>   spatial_predictor_0_40 spatial_predictor_0_41 spatial_predictor_0_42
#> 1           -0.020744827            0.063465665           -0.001583198
#> 2            0.012513154            0.006773117            0.005153518
#> 3            0.117669489            0.028793256            0.002015785
#> 4           -0.004805407           -0.052984447           -0.011659562
#> 5            0.034927798           -0.171082675            0.006518095
#> 6           -0.013496439           -0.019748363            0.002035536
#>   spatial_predictor_0_43 spatial_predictor_0_44 spatial_predictor_0_45
#> 1           0.0008763528           -0.004924952            0.004441966
#> 2          -0.0061488312            0.005445949            0.004192209
#> 3           0.0145429165            0.054685966           -0.031594959
#> 4           0.0007896651            0.012327561           -0.125926338
#> 5           0.0029789094            0.004475698           -0.086190522
#> 6           0.0001205283           -0.003080538           -0.102438242
#>   spatial_predictor_0_46 spatial_predictor_0_47 spatial_predictor_0_48
#> 1           -0.010614314            0.109509260          -0.0291230983
#> 2           -0.008707069            0.004100545          -0.0083583731
#> 3            0.154873292            0.012040765          -0.0001406001
#> 4           -0.008602096           -0.251552263           0.1535405874
#> 5           -0.026927424            0.054298969           0.0705021721
#> 6           -0.027814286            0.034531636          -0.0040461329
#>   spatial_predictor_0_49 spatial_predictor_0_50 spatial_predictor_0_51
#> 1           -0.014141207            0.065556939            0.008376850
#> 2            0.004919289            0.002300612           -0.051908297
#> 3           -0.020129602           -0.006858614            0.007107829
#> 4            0.034783817           -0.001360919           -0.012019909
#> 5            0.001021822           -0.206500175           -0.033419293
#> 6           -0.008599773           -0.018817582            0.010105730
#>   spatial_predictor_0_52 spatial_predictor_0_53 spatial_predictor_0_54
#> 1           -0.270094108            0.016737186          -0.0008228533
#> 2           -0.017095345           -0.030601816          -0.0067216761
#> 3           -0.004057733           -0.033720564           0.0245429177
#> 4           -0.203816105            0.006645515           0.0057952190
#> 5           -0.008175125           -0.005998713           0.0025015180
#> 6           -0.113070059            0.004311746           0.0054377734
#>   spatial_predictor_0_55 spatial_predictor_0_56 spatial_predictor_0_57
#> 1            0.015664149           -0.029544538            -0.08246695
#> 2           -0.031129552            0.054424622             0.02365053
#> 3           -0.211044319           -0.115905532             0.01459723
#> 4            0.016373143           -0.030249584             0.03505309
#> 5           -0.003434644           -0.014103637            -0.01536707
#> 6            0.010990643           -0.004441768             0.11325329
#>   spatial_predictor_0_58 spatial_predictor_0_59 spatial_predictor_0_60
#> 1            0.037100611            0.096731288            0.005897200
#> 2            0.044103974           -0.007519343            0.030608519
#> 3           -0.016563181            0.010981383            0.050403472
#> 4            0.002291122            0.068845992           -0.004341585
#> 5           -0.035271720           -0.122375412            0.002701816
#> 6           -0.043342941           -0.007111342           -0.005799743
#>   spatial_predictor_0_61 spatial_predictor_0_62 spatial_predictor_0_63
#> 1            0.008479632            -0.09335058           -0.077496495
#> 2           -0.098699533            -0.07575240            0.105276944
#> 3           -0.007539959            -0.01065693            0.008150333
#> 4            0.024917090             0.03804528            0.019850232
#> 5           -0.010265142             0.07225392            0.040042615
#> 6            0.029787167             0.02760241           -0.138702962
#>   spatial_predictor_0_64 spatial_predictor_0_65 spatial_predictor_0_66
#> 1           -0.011023456            0.013749019            -0.00670314
#> 2            0.004368849           -0.026456375            -0.06720143
#> 3           -0.042359390            0.036145409            -0.02388848
#> 4            0.009653611            0.003252935             0.03549700
#> 5           -0.020165538           -0.041935828            -0.01486902
#> 6           -0.057357348           -0.169761130            -0.23725453
#>   spatial_predictor_0_67 spatial_predictor_0_68 spatial_predictor_0_69
#> 1           -0.092711212             0.04605361             0.14007779
#> 2            0.026432216            -0.04457336             0.05193081
#> 3           -0.002307292             0.05047973            -0.02176933
#> 4            0.047805935             0.01809031             0.01189454
#> 5            0.086121023            -0.07767593            -0.12162162
#> 6            0.012066368            -0.08022491            -0.20666531
#>   spatial_predictor_0_70 spatial_predictor_0_71 spatial_predictor_0_72
#> 1            -0.03865851             0.01149307             0.06794522
#> 2             0.03527919             0.04203337             0.01282620
#> 3             0.05057488            -0.02103948            -0.01111083
#> 4            -0.09374814            -0.19466068             0.02195679
#> 5             0.06707658             0.05352161            -0.03062204
#> 6             0.08652559             0.01833518             0.15627827
#>   spatial_predictor_0_73 spatial_predictor_0_74 spatial_predictor_0_75
#> 1           -0.035343901           -0.017915068            0.015542313
#> 2           -0.009111780            0.005375863            0.017540288
#> 3           -0.009049244            0.036453627            0.233927322
#> 4            0.063050823            0.063122301            0.030319256
#> 5           -0.047062419           -0.031912283            0.023691678
#> 6            0.016779720            0.013204202            0.005625772
#>   spatial_predictor_0_76 spatial_predictor_0_77 spatial_predictor_0_78
#> 1             0.03292822            0.069727084           -0.017757708
#> 2            -0.03399917           -0.005818549           -0.009577624
#> 3             0.04447370           -0.032276461           -0.003526079
#> 4            -0.09559430            0.063758722            0.064366459
#> 5             0.08129817            0.182586145            0.077074235
#> 6            -0.08945159           -0.169983941            0.035904400
#>   spatial_predictor_0_79 spatial_predictor_0_80 spatial_predictor_0_81
#> 1            0.013920614            -0.00821663           -0.015901522
#> 2            0.034844339             0.01212693           -0.003600061
#> 3            0.062467634             0.03735965           -0.087326353
#> 4            0.087933528            -0.04193539           -0.041426923
#> 5            0.049456016             0.18633301            0.023410926
#> 6            0.005426446             0.05857672           -0.020355821
#>   spatial_predictor_0_82 spatial_predictor_0_83 spatial_predictor_0_84
#> 1             0.02664505            0.051483998           -0.059455466
#> 2             0.02960288            0.018758921            0.013363577
#> 3            -0.03510723           -0.009796530           -0.061048344
#> 4             0.08514307            0.054298933            0.005360485
#> 5             0.04332644           -0.044720417           -0.006508432
#> 6             0.03165506           -0.006830768           -0.034532156
#>   spatial_predictor_0_85 spatial_predictor_0_86 spatial_predictor_0_87
#> 1             0.09722934           -0.097284331            0.086610223
#> 2             0.01470533            0.003332324           -0.020797134
#> 3             0.04897573            0.034998512           -0.040119569
#> 4            -0.01867608           -0.022669681            0.007774045
#> 5            -0.01168749           -0.030586171           -0.119438460
#> 6            -0.01424540           -0.044059874           -0.104983179
#>   spatial_predictor_1000_1 spatial_predictor_1000_2 spatial_predictor_1000_3
#> 1               0.08658532              -0.03231611              0.068231938
#> 2               0.02042442              -0.15314615              0.013698668
#> 3              -0.08057675               0.07390865              0.071567363
#> 4               0.06696771              -0.06176788              0.075926333
#> 5               0.10024445               0.04832441             -0.011952038
#> 6               0.08412779               0.02766366              0.005880981
#>   spatial_predictor_1000_4 spatial_predictor_1000_5 spatial_predictor_1000_6
#> 1             3.217489e-03              -0.07018759             -0.026959299
#> 2            -2.473432e-02               0.11286917             -0.003267409
#> 3            -5.019297e-02              -0.02021901             -0.015663412
#> 4            -3.161629e-05              -0.03105710             -0.063646355
#> 5            -3.334773e-03              -0.04076160             -0.003326743
#> 6            -1.666075e-03              -0.04259223             -0.035163054
#>   spatial_predictor_1000_7 spatial_predictor_1000_8 spatial_predictor_1000_9
#> 1             0.0215975143             0.0790860856              0.005040687
#> 2            -0.0026196641            -0.0536572424             -0.005489766
#> 3             0.0162036065            -0.0605032737             -0.103856648
#> 4            -0.0009505308            -0.0007683611              0.020450989
#> 5             0.0328064995             0.1410146912             -0.101773566
#> 6             0.0340545764             0.1560500779             -0.106090804
#>   spatial_predictor_1000_10 spatial_predictor_1000_11 spatial_predictor_1000_12
#> 1               0.003350271                0.07608810               -0.03150507
#> 2              -0.046485866               -0.01337414                0.04083626
#> 3              -0.075473384                0.05253653               -0.04096583
#> 4              -0.035159892               -0.08299955               -0.11650787
#> 5               0.026591665                0.09650615                0.11436048
#> 6               0.016861260                0.03868709                0.06749009
#>   spatial_predictor_1000_13 spatial_predictor_1000_14 spatial_predictor_1000_15
#> 1               -0.07222673                0.04009140              -0.054078382
#> 2               -0.03145447               -0.08060326               0.025209320
#> 3               -0.04615498               -0.09056808               0.004550364
#> 4                0.06517494                0.01875903              -0.099293633
#> 5               -0.05184261                0.06171003               0.052480136
#> 6               -0.01316298                0.01867532              -0.005327623
#>   spatial_predictor_1000_16 spatial_predictor_1000_17 spatial_predictor_1000_18
#> 1               0.017810969              0.0113269073              -0.010257510
#> 2               0.028927460              0.0477794844               0.114391523
#> 3              -0.003050005             -0.0498092053               0.080532853
#> 4              -0.027069687             -0.0398669021               0.074806234
#> 5              -0.038380623              0.0002771587               0.032738537
#> 6               0.071443001              0.0495003413              -0.009328155
#>   spatial_predictor_1000_19 spatial_predictor_1000_20 spatial_predictor_1000_21
#> 1                0.04813674             -0.0437974692                0.01102296
#> 2               -0.03764689             -0.0183330366               -0.03000959
#> 3                0.07959245             -0.0712543891               -0.03850909
#> 4               -0.02019075              0.0099556770               -0.04529636
#> 5                0.03940559              0.0004961722               -0.02981489
#> 6                0.03468993             -0.0032133431                0.03370076
#>   spatial_predictor_1000_22 spatial_predictor_1000_23 spatial_predictor_1000_24
#> 1              -0.027440375              -0.034200698               -0.13298626
#> 2               0.218439020              -0.187647453               -0.20219303
#> 3              -0.090324377               0.073754536               -0.04432696
#> 4               0.022049436               0.003090971               -0.03809253
#> 5               0.006696832               0.023545706               -0.03388560
#> 6              -0.040124541              -0.008380628               -0.01783676
#>   spatial_predictor_1000_25 spatial_predictor_1000_26 spatial_predictor_1000_27
#> 1                0.11636893               0.132667519               0.151602883
#> 2               -0.02570768              -0.084066858              -0.040487923
#> 3               -0.10873386               0.003123965              -0.032845800
#> 4                0.02217474               0.061350239               0.002773503
#> 5                0.04045397               0.055444249              -0.067371147
#> 6                0.01030714               0.007858013               0.103901541
#>   spatial_predictor_1000_28 spatial_predictor_1000_29 spatial_predictor_1000_30
#> 1                0.12106686              0.0317377071             -0.1594341401
#> 2                0.03077180              0.0822242624             -0.0252894818
#> 3                0.01742517              0.0005693548              0.0002860974
#> 4                0.04540237              0.0294925171             -0.0148152924
#> 5               -0.05796736             -0.0203579338              0.1131264022
#> 6                0.05061757              0.0118455318             -0.1125735350
#>   spatial_predictor_1000_31 spatial_predictor_1000_32 spatial_predictor_1000_33
#> 1               -0.04135196               -0.04034484                0.03137691
#> 2                0.07494952                0.03754340               -0.05337929
#> 3                0.02038282               -0.01248751               -0.11382103
#> 4                0.13600669                0.02155827                0.06537506
#> 5                0.04101160                0.09961856               -0.02509947
#> 6               -0.13869427               -0.11630926                0.02115162
#>   spatial_predictor_1000_34 spatial_predictor_1000_35 spatial_predictor_1000_36
#> 1               -0.01452556                0.13409851               -0.07015630
#> 2               -0.08013209                0.06133915               -0.04826374
#> 3               -0.02198527               -0.03069170                0.03953152
#> 4               -0.05692688                0.11252861                0.09945197
#> 5                0.02887879                0.17843541                0.03495999
#> 6                0.05676607                0.03817728                0.03808089
#>   spatial_predictor_1000_37 spatial_predictor_1000_38 spatial_predictor_1000_39
#> 1               -0.16441816               -0.15123087               -0.01569234
#> 2                0.04761633                0.06699646                0.04794570
#> 3                0.02162981               -0.02815115                0.01841747
#> 4                0.01412370                0.06762155                0.15921531
#> 5               -0.02014062                0.20978338               -0.17502231
#> 6               -0.03451317               -0.06725425               -0.25219520
#>   spatial_predictor_1000_40 spatial_predictor_1000_41 spatial_predictor_1000_42
#> 1              -0.052881697               0.041360648               -0.03712966
#> 2              -0.045060093               0.080340444               -0.06707884
#> 3              -0.036774983               0.063167729               -0.08381353
#> 4               0.012532971               0.002011998                0.06939317
#> 5               0.001508912              -0.045524768                0.02468731
#> 6              -0.032933814               0.014634112               -0.09950179
#>   spatial_predictor_1000_43 spatial_predictor_1000_44 spatial_predictor_1000_45
#> 1              0.0009924187                0.03406906              -0.063176151
#> 2              0.0475707151               -0.06485353              -0.029631986
#> 3             -0.0795885376               -0.09240317              -0.109308147
#> 4             -0.0707769709                0.03382653               0.012411380
#> 5              0.0608986522               -0.08283213              -0.008718039
#> 6             -0.0270807705               -0.06530144               0.117763752
#>   spatial_predictor_1000_46 spatial_predictor_1000_47 spatial_predictor_1000_48
#> 1              -0.002894634                0.04297005              -0.050901484
#> 2              -0.016717014                0.02501987               0.032720687
#> 3              -0.132058215               -0.07947253               0.009571893
#> 4               0.065341767               -0.01786437               0.038196552
#> 5              -0.073750376               -0.03367780              -0.144606160
#> 6              -0.038034542                0.03967850               0.138418127
#>   spatial_predictor_1000_49 spatial_predictor_1000_50 spatial_predictor_1000_51
#> 1                0.14359592                0.01846099                0.10635259
#> 2               -0.03772670               -0.02690010               -0.05430504
#> 3                0.02445001               -0.13600703                0.02110699
#> 4               -0.03080022                0.06074740                0.14563081
#> 5                0.05615738                0.01735644               -0.02263605
#> 6               -0.16474718               -0.07841596               -0.06888592
#>   spatial_predictor_1000_52 spatial_predictor_1000_53 spatial_predictor_1000_54
#> 1              -0.000241508               -0.10875651                0.01367290
#> 2               0.015850104               -0.08289016               -0.03913177
#> 3              -0.093901157                0.05106307               -0.01105852
#> 4              -0.076717763               -0.15232997               -0.01537210
#> 5              -0.015168571               -0.01763253               -0.01283656
#> 6              -0.011287484                0.06269636                0.01206870
#>   spatial_predictor_1000_55 spatial_predictor_1000_56 spatial_predictor_1000_57
#> 1               -0.08309078              -0.032753222                0.03403235
#> 2                0.02253975               0.001842601               -0.02742228
#> 3                0.24374902               0.033593021                0.31853032
#> 4               -0.05601574              -0.008248447                0.04937029
#> 5               -0.05061777              -0.035996369                0.06950188
#> 6                0.06594581               0.078808216               -0.09638643
#>   spatial_predictor_1000_58 spatial_predictor_1000_59 spatial_predictor_1000_60
#> 1               -0.06191693                0.04216419             -0.0231419456
#> 2                0.02030021               -0.08578727              0.0000734037
#> 3               -0.03935156                0.22351219             -0.0436377191
#> 4               -0.02356013                0.03949120             -0.0371362255
#> 5                0.14295371               -0.01720984             -0.0278205957
#> 6               -0.04028605               -0.02803037             -0.0027459329
#>   spatial_predictor_1000_61 spatial_predictor_1000_62 spatial_predictor_1000_63
#> 1               0.098280214               0.079847623               -0.07078457
#> 2               0.022023086              -0.073366242               -0.07414266
#> 3               0.035103433              -0.041340942               -0.10674560
#> 4               0.012275864               0.046810287                0.10887253
#> 5               0.004667341              -0.017673312                0.07930793
#> 6               0.002605628              -0.003060575               -0.11106312
#>   spatial_predictor_1000_64 spatial_predictor_1000_65 spatial_predictor_1000_66
#> 1              -0.009341237                0.09502692               0.026757354
#> 2              -0.032701838                0.08127296              -0.031106869
#> 3              -0.047211284                0.02020700              -0.082326152
#> 4              -0.032450026               -0.04638701              -0.083501222
#> 5              -0.017794081               -0.06405042               0.003938521
#> 6              -0.032329644                0.12832392              -0.038232356
#>   spatial_predictor_1000_67 spatial_predictor_1000_68 spatial_predictor_1000_69
#> 1               0.017717805              -0.056097200                0.02021046
#> 2              -0.036041310               0.053447156                0.03501488
#> 3              -0.024005564              -0.035042736               -0.12319399
#> 4               0.057734523               0.090578300                0.00989136
#> 5               0.001937527              -0.002230929               -0.07911011
#> 6               0.094831138               0.108278776                0.14358105
#>   spatial_predictor_1000_70 spatial_predictor_1000_71 spatial_predictor_1000_72
#> 1              -0.037729770                0.05672643               -0.14152073
#> 2               0.044920980               -0.09919566               -0.10969028
#> 3              -0.069083674                0.02324144                0.01200051
#> 4              -0.069607759                0.02891570               -0.08000965
#> 5              -0.002556965               -0.07422993               -0.07798042
#> 6              -0.030013400               -0.08394401               -0.03210298
#>   spatial_predictor_1000_73 spatial_predictor_2000_1 spatial_predictor_2000_2
#> 1              -0.019293352              -0.08631926               0.02040360
#> 2              -0.039012292              -0.01354870               0.11442967
#> 3              -0.099882513               0.08362924              -0.10233266
#> 4               0.084892409              -0.05351554               0.05909536
#> 5              -0.057933974              -0.12588173              -0.06028829
#> 6              -0.002781528              -0.09043986              -0.01080571
#>   spatial_predictor_2000_3 spatial_predictor_2000_4 spatial_predictor_2000_5
#> 1              -0.06509408             -0.008431511              -0.06362221
#> 2              -0.03332258              0.025772123               0.03445967
#> 3              -0.08158425              0.039493911               0.01413149
#> 4              -0.08904390             -0.006217016               0.02879223
#> 5               0.03660447              0.001568544              -0.06327706
#> 6              -0.03340558             -0.006621275              -0.04942669
#>   spatial_predictor_2000_6 spatial_predictor_2000_7 spatial_predictor_2000_8
#> 1              0.048301791               0.02152755               0.06365469
#> 2             -0.018383801               0.03509728               0.03334159
#> 3             -0.020298111              -0.02103241               0.05083017
#> 4              0.051786910              -0.02792472              -0.03826332
#> 5             -0.008811236              -0.03994246              -0.04225077
#> 6              0.073816113               0.05690363               0.12885417
#>   spatial_predictor_2000_9 spatial_predictor_2000_10 spatial_predictor_2000_11
#> 1             -0.084751874               -0.09938075               -0.05794812
#> 2             -0.004656693                0.04826282                0.08428799
#> 3             -0.007460976                0.05113736                0.07422658
#> 4              0.067450117                0.04757035                0.01467080
#> 5              0.071143678                0.18415663                0.01284579
#> 6             -0.115630321               -0.04977177               -0.09561849
#>   spatial_predictor_2000_12 spatial_predictor_2000_13 spatial_predictor_2000_14
#> 1                0.07457584              -0.002455888               -0.00799680
#> 2               -0.03167288              -0.075605529                0.09881847
#> 3                0.01328955               0.004448900               -0.06276480
#> 4                0.04015331              -0.028564751                0.02322024
#> 5               -0.12676537               0.028114077                0.02049014
#> 6               -0.06060799               0.017592450                0.06271532
#>   spatial_predictor_2000_15 spatial_predictor_2000_16 spatial_predictor_2000_17
#> 1                0.13427751               -0.02487087              -0.007307794
#> 2                0.02953363               -0.08579860               0.027361090
#> 3                0.02482007               -0.02522371               0.024303903
#> 4                0.09233785                0.07089192              -0.004536872
#> 5               -0.04674877               -0.04280002              -0.127031756
#> 6                0.11443744                0.19608469               0.044679646
#>   spatial_predictor_2000_18 spatial_predictor_2000_19 spatial_predictor_2000_20
#> 1              -0.101330037               -0.03359234               0.063516738
#> 2               0.024709929               -0.04041777               0.086290047
#> 3              -0.034442553                0.04003291              -0.030245913
#> 4               0.005434962               -0.01271874              -0.065533741
#> 5               0.038747462               -0.02955281              -0.060507804
#> 6              -0.039877064                0.20569228              -0.006294619
#>   spatial_predictor_2000_21 spatial_predictor_2000_22 spatial_predictor_2000_23
#> 1             -0.1546135782              0.0054285399               -0.11379109
#> 2              0.0612927733             -0.0642446743               -0.07582007
#> 3              0.0232584718             -0.0009050286               -0.07544447
#> 4              0.0489870703              0.0116493336                0.05020771
#> 5              0.0144215263              0.0092313010                0.01544571
#> 6              0.0005940138              0.0419692678               -0.01909092
#>   spatial_predictor_2000_24 spatial_predictor_2000_25 spatial_predictor_2000_26
#> 1              -0.067317935               -0.13263348               -0.04939068
#> 2               0.109823340                0.01333682               -0.05519138
#> 3              -0.072956910                0.01571004                0.05121848
#> 4              -0.001519644               -0.04826864               -0.04933506
#> 5              -0.041505520               -0.08087331               -0.12417977
#> 6              -0.082534372               -0.13716319               -0.15128073
#>   spatial_predictor_2000_27 spatial_predictor_2000_28 spatial_predictor_2000_29
#> 1               -0.04169996               -0.05133889                0.08501408
#> 2                0.08267045                0.24190162               -0.11484149
#> 3               -0.08266907                0.00397999               -0.06209341
#> 4               -0.01785186                0.01458162               -0.03479379
#> 5               -0.09228334                0.02695580                0.07162798
#> 6               -0.06586184               -0.11484956                0.06999691
#>   spatial_predictor_2000_30 spatial_predictor_2000_31 spatial_predictor_2000_32
#> 1                0.01203364              -0.007903691                0.03647948
#> 2               -0.01864633              -0.062705431                0.05883284
#> 3               -0.07611588              -0.089497674                0.03595028
#> 4                0.08001896               0.052770832                0.06881429
#> 5                0.02011409               0.110097491                0.06164528
#> 6                0.02823743               0.034885144                0.04464014
#>   spatial_predictor_2000_33 spatial_predictor_2000_34 spatial_predictor_2000_35
#> 1              -0.053450430               0.042119183               0.018017978
#> 2               0.002369092               0.001515834               0.022633636
#> 3              -0.003430045               0.003561290               0.143201816
#> 4               0.061887004               0.069674924              -0.007572006
#> 5              -0.101880397               0.078109408              -0.028985924
#> 6              -0.023728867              -0.020781870               0.095350631
#>   spatial_predictor_2000_36 spatial_predictor_2000_37 spatial_predictor_2000_38
#> 1              -0.085610937              -0.084913948               0.114177908
#> 2              -0.074256973              -0.036871539               0.048749014
#> 3               0.056376439               0.004168657               0.163207894
#> 4              -0.008521845              -0.067818590               0.017338699
#> 5               0.002444705               0.039586831               0.009668638
#> 6              -0.042534715              -0.046493702               0.030815493
#>   spatial_predictor_2000_39 spatial_predictor_2000_40 spatial_predictor_2000_41
#> 1               0.015755092               -0.07839766              -0.101238720
#> 2              -0.000243810                0.09107001               0.081581481
#> 3               0.148560892                0.09433258               0.001729471
#> 4               0.082438305               -0.05777713               0.027249717
#> 5              -0.009087817                0.01543621              -0.047145598
#> 6              -0.059245745                0.01733381              -0.066636163
#>   spatial_predictor_2000_42 spatial_predictor_2000_43 spatial_predictor_2000_44
#> 1               -0.09609810                0.07460715               -0.07317322
#> 2               -0.03185668               -0.01841866               -0.03654769
#> 3                0.05667748                0.08197819               -0.02469837
#> 4               -0.05788700               -0.07489249                0.01719729
#> 5                0.12416273                0.04699518               -0.06028196
#> 6                0.02242201                0.01573111               -0.14193308
#>   spatial_predictor_2000_45 spatial_predictor_2000_46 spatial_predictor_2000_47
#> 1              -0.033256754                0.10636699               -0.09812679
#> 2              -0.095326933                0.10934097               -0.01877419
#> 3              -0.016383460               -0.13071471               -0.05262413
#> 4              -0.021046662               -0.11727764               -0.03930079
#> 5               0.052992077               -0.04276183                0.02582947
#> 6              -0.008921881                0.01430250               -0.02620040
#>   spatial_predictor_2000_48 spatial_predictor_2000_49 spatial_predictor_2000_50
#> 1               -0.09175040               0.077160844               0.049278646
#> 2                0.07171080              -0.014493695               0.020306032
#> 3               -0.02203282              -0.041213158              -0.058290532
#> 4                0.02563489               0.002433778              -0.007742549
#> 5                0.07421646               0.059236959              -0.079239541
#> 6               -0.01438554               0.032590728              -0.065926013
#>   spatial_predictor_2000_51 spatial_predictor_2000_52 spatial_predictor_2000_53
#> 1               -0.01379994               0.012797830              -0.015611160
#> 2               -0.02818744               0.098782855              -0.004278481
#> 3               -0.01733597               0.042303948              -0.040424639
#> 4               -0.02820731              -0.005720328               0.018174463
#> 5               -0.02050898              -0.072008192              -0.028520555
#> 6                0.03106434              -0.064126109              -0.056606435
#>   spatial_predictor_2000_54 spatial_predictor_2000_55 spatial_predictor_2000_56
#> 1              0.0779580223              -0.030450995                0.06215135
#> 2              0.1191718376              -0.001923934                0.10540988
#> 3             -0.0621142410              -0.038813839               -0.09497336
#> 4              0.0466335027              -0.017615998                0.02959161
#> 5              0.0727755038              -0.006342720                0.03699486
#> 6              0.0002328952              -0.038641102                0.02084605
#>   spatial_predictor_2000_57 spatial_predictor_2000_58 spatial_predictor_2000_59
#> 1               -0.13452386                0.01034527                0.12058833
#> 2               -0.02834579                0.07638533               -0.05340681
#> 3               -0.02696148               -0.01931001                0.02038225
#> 4               -0.02716742               -0.01555246                0.03723406
#> 5                0.10713255                0.04030510               -0.05531802
#> 6                0.08458113                0.03305808                0.01834042
#>   spatial_predictor_2000_60 spatial_predictor_2000_61 spatial_predictor_2000_62
#> 1               -0.08282613                0.02918915                0.05893357
#> 2               -0.12173382               -0.05741090                0.01473379
#> 3                0.12008004               -0.06150843               -0.01837377
#> 4               -0.09272016               -0.02522304               -0.06222357
#> 5               -0.05507015               -0.04800251                0.03463085
#> 6               -0.04511586                0.05624974                0.05356797
#>   spatial_predictor_2000_63 spatial_predictor_2000_64 spatial_predictor_2000_65
#> 1              -0.044184426                0.18469458                0.04367228
#> 2               0.027767409               -0.06619654                0.16827215
#> 3              -0.090851592                0.07404111               -0.04019892
#> 4              -0.050980670               -0.03931814               -0.10198484
#> 5              -0.078543142               -0.07977692               -0.02877802
#> 6              -0.008705952               -0.15131150               -0.03108294
#>   spatial_predictor_2000_66 spatial_predictor_2000_67 spatial_predictor_2000_68
#> 1                0.05531886              -0.036439661               -0.04486743
#> 2               -0.02417932               0.162485264                0.02394099
#> 3                0.01952697               0.070056200               -0.10106037
#> 4                0.05700922               0.004868591                0.09437313
#> 5                0.09349474               0.069183976                0.01947779
#> 6                0.03294707               0.075260018                0.02777286
#>   spatial_predictor_2000_69 spatial_predictor_2000_70 spatial_predictor_2000_71
#> 1                0.13190182               0.039575210               -0.02655051
#> 2                0.03474420              -0.003326546                0.10475583
#> 3               -0.06074965              -0.154220879                0.05036135
#> 4               -0.02739143              -0.048871809                0.13973483
#> 5                0.00555685              -0.041575483               -0.06676613
#> 6               -0.02899025              -0.095451088                0.06302251
#>   spatial_predictor_2000_72 spatial_predictor_2000_73 spatial_predictor_2000_74
#> 1               0.119902705                0.05530457               -0.07687582
#> 2               0.046308251               -0.07474417                0.06063789
#> 3              -0.023066581               -0.04585847                0.01918167
#> 4               0.007382971                0.02230235               -0.01940659
#> 5              -0.015930377               -0.03557995                0.02183616
#> 6               0.012060844               -0.10011422                0.04739226