Normality test of a numeric vector

Applies a Shapiro-Wilks test to a numeric vector, and returns a list with the statistic W, its p-value, and a character string with the interpretation.

Usage

residuals_test(residuals)

Arguments

residuals

Numeric vector, model residuals.

Value

A list with four slots:

/item w W statistic returned by shapiro.test(). /item p.value p-value of the Shapiro test. /item interpretation Character vector, one of "x is normal", "x is not normal". /item plot A patchwork plot with the qq plot and the histogram of x.

See also

Examples

residuals_test(residuals = runif(100))
#> $shapiro.w
#> [1] 0.9606738
#> 
#> $p.value
#> [1] 0.004492854
#> 
#> $interpretation
#> [1] "Residuals are not normal"
#>