Plots the response curves of a model.

Plots the response curves of models fitted with rf(), rf_repeat(), or rf_spatial().

plot_response_curves(
  x = NULL,
  variables = NULL,
  quantiles = c(0.1, 0.5, 0.9),
  grid.resolution = 200,
  ncol = NULL,
  show.data = FALSE,
  verbose = TRUE
)

Arguments

x
A model fitted with rf(), rf_repeat(), or rf_spatial().

variables
Character vector, names of predictors to plot. If NULL, the most important variables (importance higher than the median) in x are selected. Default: NULL.

quantiles
Numeric vector with values between 0 and 1, argument probs of quantile. Quantiles to set the other variables to. Default: c(0.1, 0.5, 0.9)

grid.resolution
Integer between 20 and 500. Resolution of the plotted curve Default: 100

ncol
Integer, argument of wrap_plots. Defaults to the rounded squared root of the number of plots.

show.data
Logical, if TRUE, the observed data is plotted along with the response curves. Default. FALSE

verbose
Logical, if TRUE the plot is printed. Default: TRUE

Value

A list with slots named after the selected variables, with one ggplot each.

Details

All variables that are not plotted in a particular response curve are set to the values of their respective quantiles, and the response curve for each one of these quantiles is shown in the plot. When the input model was fitted with rf_repeat() with keep.models = TRUE, then the plot shows the median of all model runs, and each model run separately as a thinner line. The output list can be plotted all at once with patchwork::wrap_plots(p) or cowplot::plot_grid(plotlist = p), or one by one by extracting each plot from the list.

See also

Examples

# \donttest{
if(interactive()){
data(plant_richness_df)

out <- rf(
 data = plant_richness_df,
 dependent.variable.name = "richness_species_vascular",
 predictor.variable.names = colnames(plant_richness_df)[5:21],
 verbose = FALSE
)

p <- plot_response_curves(x = out)
p

}
# }