print_performance

Prints the performance slot of a model fitted with rf(), rf_repeat(), or rf_spatial(). For models fitted with rf_repeat() it shows the median and the median absolute deviation of each performance measure.

Usage

print_performance(model)

Arguments

model

Model fitted with rf(), rf_repeat(), or rf_spatial().

Value

Prints model performance scores to the console.

See also

Examples

data(plants_rf)

print_performance(plants_rf)
#> 
#> Model performance 
#>   - R squared (oob):                  0.5015626
#>   - R squared (cor(obs, pred)^2):     0.8378728
#>   - Pseudo R squared (cor(obs, pred)):0.9153539
#>   - RMSE (oob):                       2379.254
#>   - RMSE:                             1609.614
#>   - Normalized RMSE:                  0.4646691