Standard error of the mean of a numeric vector

Computes the standard error of the mean of a numeric vector as round(sqrt(var(x)/length(x)), 3)

Usage

standard_error(x)

Arguments

x

A numeric vector.

Value

A numeric value.

Details

The function removes NA values before computing the standard error, and rounds the result to 3 decimal places.

See also

Examples

standard_error(x = runif(10))
#> [1] 0.071