Multicollinearity reduction via Pearson correlation

Computes the correlation matrix among a set of predictors, orders the correlation matrix according to a user-defined preference order, and removes variables one by one, taking into account the preference order, until the remaining ones are below a given Pearson correlation threshold. Warning: variables in preference.order not in colnames(x), columns with zero variance, and non-numeric columns are removed silently from x and preference.order. The same happens with rows having NA values (na.omit() is applied).

auto_cor(
  x = NULL,
  preference.order = NULL,
  cor.threshold = 0.75,
  verbose = TRUE
)

Arguments

x
A data frame with predictors, or the result of auto_vif() Default: NULL.

preference.order
Character vector indicating the user's order of preference to keep variables. Doesn't need to contain If not provided, variables in x are prioritised by their column order. Default: NULL.

cor.threshold
Numeric between 0 and 1, with recommended values between 0.5 and 0.9. Maximum Pearson correlation between any pair of the selected variables. Default: 0.75

verbose
Logical. if TRUE, describes the function operations to the user. Default:: TRUE

Value

List with three slots:

cor: correlation matrix of the selected variables.

selected.variables: character vector with the names of the selected variables.

selected.variables.df: data frame with the selected variables.

Details

Can be chained together with auto_vif() through pipes, see the examples below.

See also

Examples

# \donttest{
if(interactive()){

 data(plant_richness_df)

 #on a data frame
 out <- auto_cor(x = plant_richness_df[, 4:21])
 out$selected.variables

 #on the result of auto_vif
 out <- auto_vif(x = plant_richness_df[, 5:20])
 out <- auto_cor(x = out)

 #with pipes (cor and vif thresholds are arbitrary)
 out <- plant_richness_df[, 5:20] %>%
 auto_vif(vif.threshold = 2.5) %>%
 auto_cor(cor.threshold = 0.4)

 }
# }