Makes one training and one testing spatial folds

Used internally by make_spatial_folds() and rf_evaluate(). Uses the coordinates of a point xy.i to generate two spatially independent data folds from the data frame xy. It does so by growing a rectangular buffer from xy.i until a number of records defined by training.fraction is inside the buffer. The indices of these records are then stored as "training" in the output list. The indices of the remaining records outside of the buffer are stored as "testing". These training and testing records can be then used to evaluate a model on independent data via cross-validation.

Usage

make_spatial_fold(
  data = NULL,
  dependent.variable.name = NULL,
  xy.i = NULL,
  xy = NULL,
  distance.step.x = NULL,
  distance.step.y = NULL,
  training.fraction = 0.8
)

Arguments

data

Data frame with a response variable and a set of predictors. Default: NULL

dependent.variable.name

Character string with the name of the response variable. Must be in the column names of data. Default: NULL

xy.i

One row data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, id of the record). Can be a row of xy. Default: NULL.

xy

A data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, index of the record). Default: NULL.

distance.step.x

Numeric, distance step used during the growth in the x axis of the buffers defining the training folds. Default: NULL (1/1000th the range of the x coordinates).

distance.step.y

Numeric, distance step used during the growth in the y axis of the buffers defining the training folds. Default: NULL (1/1000th the range of the y coordinates).

training.fraction

Numeric, fraction of the data to be included in the training fold, Default: 0.8.

Value

A list with two slots named training and testing with the former having the indices of the training records selected from xy, and the latter having the indices of the testing records.

See also

Examples

 #loading example data
 data(plant_richness_df)

 #getting case coordinates
 xy <- plant_richness_df[, 1:3]
 colnames(xy) <- c("id", "x", "y")

 #building a spatial fold centered in the first pair of coordinates
 out <- make_spatial_fold(
   xy.i = xy[1, ],
   xy = xy,
   training.fraction = 0.6
 )

 #indices of the training and testing folds
 out$training
#>   [1]   1   2   4   5   6   7  10  12  14  15  19  20  21  22  23  26  28  31
#>  [19]  32  33  34  35  36  37  45  47  48  50  51  54  56  58  59  62  63  64
#>  [37]  65  66  67  71  73  74  77  78  79  80  81  83  87  94  95  96  97  98
#>  [55]  99 100 102 104 106 107 108 109 110 111 112 113 114 115 118 119 122 125
#>  [73] 126 127 129 131 135 136 138 139 141 145 151 153 154 155 156 157 159 160
#>  [91] 161 162 163 165 166 168 171 174 177 178 179 181 182 185 186 187 188 189
#> [109] 191 192 193 194 195 198 201 204 205 206 207 208 209 210 212 213 214 215
#> [127] 216 218 219 220 221 222 223 224 226 227
 out$testing
#>  [1]   3   8   9  11  13  16  17  18  24  25  27  29  30  38  39  40  41  42  43
#> [20]  44  46  49  52  53  55  57  60  61  68  69  70  72  75  76  82  84  85  86
#> [39]  88  89  90  91  92  93 101 103 105 116 117 120 121 123 124 128 130 132 133
#> [58] 134 137 140 142 143 144 146 147 148 149 150 152 158 164 167 169 170 172 173
#> [77] 175 176 180 183 184 190 196 197 199 200 202 203 211 217 225

if(interactive()){

 #plotting the data
 plot(xy[ c("x", "y")], type = "n", xlab = "", ylab = "")
 #plots training points
 points(xy[out$training, c("x", "y")], col = "red4", pch = 15)
 #plots testing points
 points(xy[out$testing, c("x", "y")], col = "blue4", pch = 15)
 #plots xy.i
 points(xy[1, c("x", "y")], col = "black", pch = 15, cex = 2)

}