Makes one training and one testing spatial folds

Used internally by make_spatial_folds() and rf_evaluate(). Uses the coordinates of a point xy.i to generate two spatially independent data folds from the data frame xy. It does so by growing a rectangular buffer from xy.i until a number of records defined by training.fraction is inside the buffer. The indices of these records are then stored as "training" in the output list. The indices of the remaining records outside of the buffer are stored as "testing". These training and testing records can be then used to evaluate a model on independent data via cross-validation.

make_spatial_fold(
  xy.i = NULL,
  xy = NULL,
  distance.step = NULL,
  training.fraction = 0.6
)

Arguments

xy.i
One row data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, id of the record). Can be a row of xy. Default: NULL.

xy
A data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, index of the record). Default: NULL.

distance.step
Numeric, distance step used during the thinning iterations. If NULL, the one percent of the maximum distance among points in xy is used. Default: NULL

training.fraction
Numeric, fraction of the data to be included in the training fold, Default: 0.6.

Value

A list with two slots named training and testing with the former having the indices of the training records selected from xy, and the latter having the indices of the testing records.

See also

Examples

# \donttest{
if(interactive()){

 data(plant_richness_df)

 xy <- plant_richness_df[, 1:3]
 colnames(xy) <- c("id", "x", "y")
 out <- make_spatial_fold(
   xy.i = xy[1, ],
   xy = xy,
   distance.step = 0.05, #degrees
   training.fraction = 0.6
 )

 out$training
 out$testing

 plot(xy[ c("x", "y")], type = "n", xlab = "", ylab = "")
 #plots training points
 points(xy[out$training, c("x", "y")], col = "red4", pch = 15)
 #plots testing points
 points(xy[out$testing, c("x", "y")], col = "blue4", pch = 15)
 #plots xy.i
 points(xy[1, c("x", "y")], col = "black", pch = 15, cex = 2)

}
# }