Makes training and testing spatial folds

Applies make_spatial_fold() to every record in a data frame xy.selected to generate as many spatially independent folds over the dataset xy as rows are in xy.selected.

make_spatial_folds(
  data = NULL,
  dependent.variable.name = NULL,
  xy.selected = NULL,
  xy = NULL,
  distance.step = NULL,
  training.fraction = 0.6,
  n.cores = NULL
)

Arguments

data
Data frame with a response variable and a set of predictors. Default: NULL

dependent.variable.name
Character string with the name of the response variable. Must be in the column names of data. Default: NULL

xy.selected
Data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, id of the record). Usually a subset of xy. Usually the result of applying thinning() or thinning_til_n() to 'xy' Default: NULL.

xy
data frame with at least three columns: "x" (longitude), "y" (latitude), and "id" (integer, index of the record). Default: NULL.

distance.step
Numeric, distance step used during the thinning iterations. If NULL, the maximum distance between two points in xy divided by 100 is used. Default: NULL

training.fraction
numeric, fraction of the data to be included in the growing buffer as training data, Default: 0.6

n.cores
number of cores to use to generate spatial folds in parallel. Default: NULL.

Value

A list with as many slots as rows are in xy.selected. Each slot has two slots named training and testing, with the former having the indices of the training records selected from xy, and the latter having the indices of the testing records.

See also

Examples

# \donttest{
if(interactive()){

 data(plant_richness_df)

 xy <- plant_richness_df[, 1:3]
 colnames(xy) <- c("id", "x", "y")
 xy.selected <- thinning_til_n(
   xy = xy,
   n = 20
   )

 out <- make_spatial_folds(
   xy.selected = xy.selected,
   xy = xy,
   distance.step = 0.05, #degrees
   training.fraction = 0.6,
   n.cores = 1
 )

 length(out)

 plot(xy[ c("x", "y")], type = "n", xlab = "", ylab = "")
 #plots training points
 points(xy[out[[10]]$training, c("x", "y")], col = "red4", pch = 15)
 #plots testing points
 points(xy[out[[10]]$testing, c("x", "y")], col = "blue4", pch = 15)
 #plots xy.i
 points(xy[10, c("x", "y")], col = "black", pch = 15, cex = 2)
}
# }