Moran's I test on a numeric vector

Computes the spatial correlation coefficient (Moran's I) of a vector given a distance matrix, and a distance threshold used to define neighborhood.

moran(
  x = NULL,
  distance.matrix = NULL,
  distance.threshold = 0
)

Arguments

x
Numeric vector, generally model residuals, Default: NULL

distance.matrix
Distance matrix among cases in x. The number of rows of this matrix must be equal to the length of x. Default: NULL

distance.threshold
numeric value in the range of values available in distance.matrix. Distances below such threshold are set to 0. Default: 0

Value

A list with three named slots:

moran.i: Moran's I of x.

p.value: P-value of the Moran's I estimate.

interpretation: Interpretation of the Moran's I value according to the p-value. One of "Positive spatial correlation", "Negative spatial correlation", and "No spatial correlation".

Details

Inspired in the Moran.I() function of the ape package.

See also

Examples

# \donttest{
if(interactive()){

 data(distance_matrix)
 data(plant_richness)

 moran.out <- moran(
   x = plant_richness$richness_species_vascular,
   distance.matrix = distance_matrix
   )
 moran.out

}
# }