Scatterplots of a training data frame

Plots the dependent variable against each predictor.

plot_training_df(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  ncol = 4,
  method = "loess"
)

Arguments

data
Data frame with a response variable and a set of predictors. Default: NULL

dependent.variable.name
Character string with the name of the response variable. Must be in the column names of data. If the dependent variable is binary with values 1 and 0, the argument case.weights of ranger is populated by the function case_weights(). Default: NULL

predictor.variable.names
Character vector with the names of the predictive variables. Every element of this vector must be in the column names of data. Optionally, the result of auto_cor() or auto_vif() Default: NULL

ncol
Number of columns of the plot. Argument ncol of wrap_plots.

method
Method for geom_smooth, one of: "lm", "glm", "gam", "loess", or a function, for example mgcv::gam Default: 'loess'

Value

A wrap_plots object.

Examples

if (FALSE) {
if(interactive()){
   data(plant_richness_df)
   plot_training_data(
     data = plant_richness_df,
     dependent.variable.name = "richness_species_vascular",
     predictor.variable.names = colnames(plant_richness_df)[5:21]
     )
 }
}