Ranks spatial predictors

Ranks spatial predictors generated by mem_multithreshold() or pca_multithreshold() by their effect in reducing the Moran's I of the model residuals (ranking.method = "effect"), or by their own Moran's I (ranking.method = "moran").

In the former case, one model of the type y ~ predictors + spatial_predictor_X is fitted per spatial predictor, and the Moran's I of the model residuals residuals is compared with the one of the model without spatial predictors (y ~ predictors), to finally rank the spatial predictor from maximum to minimum difference in Moran's I.

In the latter case, the spatial predictors are ordered by their Moran's I alone (this is the faster option).

In both cases, spatial predictors with no effect (no reduction of Moran's I or Moran's I of the spatial predictor equal or lower than 0) are removed.

This function has been designed to be used internally by rf_spatial() rather than directly by a user.

rank_spatial_predictors(
  data = NULL,
  dependent.variable.name = NULL,
  predictor.variable.names = NULL,
  distance.matrix = NULL,
  distance.thresholds = NULL,
  ranger.arguments = NULL,
  spatial.predictors.df = NULL,
  ranking.method = c("effect", "moran"),
  reference.moran.i = 1,
  verbose = FALSE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = "11000"
)

Arguments

data
Data frame with a response variable and a set of predictors. Default: NULL

dependent.variable.name
Character string with the name of the response variable. Must be in the column names of data. Default: NULL

predictor.variable.names
Character vector with the names of the predictive variables. Every element of this vector must be in the column names of data. Default: NULL

distance.matrix
Squared matrix with the distances among the records in data. The number of rows of distance.matrix and data must be the same. If not provided, the computation of the Moran's I of the residuals is omitted. Default: NULL

distance.thresholds
Numeric vector with neighborhood distances. All distances in the distance matrix below each value in dustance.thresholds are set to 0 for the computation of Moran's I. If NULL, it defaults to seq(0, max(distance.matrix), length.out = 4). Default: NULL

ranger.arguments
List with ranger arguments. See rf or rf_repeat for further details.

spatial.predictors.df
Data frame of spatial predictors.

ranking.method
Character, method used by to rank spatial predictors. The method "effect" ranks spatial predictors according how much each predictor reduces Moran's I of the model residuals, while the method "moran" ranks them by their own Moran's I.

reference.moran.i
Moran's I of the residuals of the model without spatial predictors. Default: 1

verbose
Logical, ff TRUE, messages and plots generated during the execution of the function are displayed, Default: TRUE

n.cores
Integer, number of cores to use during computations. If NULL, all cores but one are used, unless a cluster is used. Default = NULL

cluster.ips
Character vector with the IPs of the machines in a cluster. The machine with the first IP will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.

cluster.cores
Numeric integer vector, number of cores to use on each machine.

cluster.user
Character string, name of the user (should be the same throughout machines). Defaults to the current system user.

cluster.port
Character, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: "11000"

Value

A list with three slots:

method: Character, name of the method used to rank the spatial predictors.

criteria: Data frame with two different configurations depending on the ranking method. If ranking.method = "effect", the columns contain the names of the spatial predictors, the r-squared of the model, the Moran's I of the model residuals, the difference between the Moran's I of the model including the given spatial predictor, and the Moran's I of the model fitted without spatial predictors, and the interpretation of the Moran's I value. If ranking.method = "moran", only the name of the spatial predictor and it's Moran's I are in the output data frame.

ranking: Ordered character vector with the names of the spatial predictors selected.

Examples

# \donttest{
if(interactive()){
 data("distance_matrix")

 spatial.predictors.df <- pca_distance_matrix(
   x = distance_matrix[1:50, 1:50],
   distance.thresholds = c(0, 100, 1000)
 )

 #ranking by the Moran's I of the spatial predictor
 rank <- rank_spatial_predictors(
   distance.matrix = distance_matrix[1:50, 1:50],
   distance.thresholds = c(0, 100, 1000),
   spatial.predictors.df = spatial.predictors.df,
   ranking.method = "effect",
   n.cores = 1
 )
 rank$ranking.criteria
 rank$ranking
 }
# }