Compares the performance of several models on independent data

Uses rf_evaluate() to compare the performance of several models based on the same pairs of coordinates on independent spatial folds.

rf_compare(
  models = NULL,
  xy = NULL,
  repetitions = 30,
  training.fraction = 0.8,
  metrics = c("r.squared", "pseudo.r.squared", "rmse", "nrmse"),
  fill.color = viridis::viridis(100, option = "F", direction = -1, alpha = 0.8),
  line.color = "gray30",
  seed = NULL,
  verbose = TRUE,
  n.cores = parallel::detectCores() - 1,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = "11000"
)

Arguments

models
Named list with models based on the same pairs of coordinates. Example: models = list(a = model.a, b = model.b). Default: NULL

xy
Data frame or matrix with two columns containing coordinates and named "x" and "y". The same xy is used for all models, that's why all models must be based on the same pairs of coordinates. If NULL, the function will throw an error. Default: NULL

repetitions
Integer, must be lower than the total number of rows available in the model's data. Default: 30

training.fraction
Proportion between 0.5 and 0.9 indicating the number of records to be used in model training. Default: 0.8

metrics
Character vector, names of the performance metrics selected. The possible values are: "r.squared" (cor(obs, pred) ^ 2), "pseudo.r.squared" (cor(obs, pred)), "rmse" (sqrt(sum((obs - pred)^2)/length(obs))), "nrmse" (rmse/(quantile(obs, 0.75) - quantile(obs, 0.25))). Default: c("r.squared", "pseudo.r.squared", "rmse", "nrmse")

fill.color
Character vector with hexadecimal codes (e.g. "#440154FF" "#21908CFF" "#FDE725FF"), or function generating a palette (e.g. viridis::viridis(100)). Default: viridis::viridis(100, option = "F", direction = -1)

line.color
Character string, color of the line produced by ggplot2::geom_smooth(). Default: "gray30"

seed
Integer, random seed to facilitate reproduciblity. If set to a given number, the results of the function are always the same.

verbose
Logical. If TRUE, messages and plots generated during the execution of the function are displayed, Default: TRUE

n.cores
Integer, number of cores to use. Default = parallel::detectCores() - 1

cluster.ips
Character vector with the IPs of the machines in a cluster. The machine with the first IP will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.

cluster.cores
Numeric integer vector, number of cores to use on each machine.

cluster.user
Character string, name of the user (should be the same throughout machines). Defaults to the current system user.

cluster.port
Character, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: "11000"

Value

A list with three slots: comparison.df, a data frame with one performance value per spatial fold, metric, and model; spatial.folds, a list with the indices of the training and testing records for each evaluation repetition; plot a boxplot of comparison.df.

See also

Examples

# \donttest{
if(interactive()){

 data(distance_matrix)
 data(plant_richness_df)

 rf.model <- rf(
   data = plant_richness_df,
   dependent.variable.name = "richness_species_vascular",
   predictor.variable.names = colnames(plant_richness_df)[5:21],
   distance.matrix = distance_matrix,
   distance.thresholds = c(0, 1000, 2000)
 )

 rf.spatial <- rf_spatial(model = rf.model)

comparison <- rf_compare(
 models = list(
   `Non spatial` = rf.model,
   Spatial = rf.spatial
 ),
 xy = plant_richness_df[, c("x", "y")],
 metrics = c("r.squared", "rmse")
)

 }
# }