Tuning of random forest hyperparameters

Tunes the random forest hyperparameters num.trees, mtry, and min.node.size via grid search by maximizing the model's R squared. Two methods are available: out-of-bag (oob), and spatial cross-validation performed with rf_evaluate().

rf_tuning(
  model = NULL,
  method = c("oob", "spatial.cv"),
  num.trees = NULL,
  mtry = NULL,
  min.node.size = NULL,
  xy = NULL,
  repetitions = NULL,
  training.fraction = 0.8,
  verbose = TRUE,
  n.cores = NULL,
  cluster.ips = NULL,
  cluster.cores = NULL,
  cluster.user = Sys.info()[["user"]],
  cluster.port = "11000"
)

Arguments

model
A model fitted with rf(). If provided, the training data is taken directly from the model definition (stored in model$ranger.arguments). Default: NULL

method
Character, "oob" to use RMSE values computed on the out-of-bag data to guide the tuning, and "spatial.cv", to use RMSE values from a spatial cross-validation on independent spatial folds done via rf_evaluate(). Default: "oob"

num.trees
Numeric integer vector with the number of trees to fit on each model repetition. Default: c(500, 1000, 2000).

mtry
Numeric integer vector, number of predictors to randomly select from the complete pool of predictors on each tree split. Default: floor(seq(1, length(predictor.variable.names), length.out = 4))

min.node.size
Numeric integer, minimal number of cases in a terminal node. Default: c(5, 10, 20, 40)

xy
Data frame or matrix with two columns containing coordinates and named "x" and "y". If NULL, the function will throw an error. Default: NULL

repetitions
Integer, number of repetitions to compute the R squared from. If method = "oob", number of repetitions to be used in rf_repeat() to fit models for each combination of hyperparameters. If method = "spatial.cv", number of independent spatial folds to use during the cross-validation. Default: NULL (which yields 30 for "spatial.cv" and 5 for "oob").

training.fraction
Proportion between 0.2 and 0.8 indicating the number of records to be used in model training. Default: 0.8

verbose
Logical. If TRUE, messages and plots generated during the execution of the function are displayed, Default: TRUE

n.cores
Integer, number of cores to use during computations. If NULL, all cores but one are used, unless a cluster is used. Default = NULL

cluster.ips
Character vector with the IPs of the machines in a cluster. The machine with the first IP will be considered the main node of the cluster, and will generally be the machine on which the R code is being executed.

cluster.cores
Numeric integer vector, number of cores to use on each machine.

cluster.user
Character string, name of the user (should be the same throughout machines). Defaults to the current system user.

cluster.port
Character, port used by the machines in the cluster to communicate. The firewall in all computers must allow traffic from and to such port. Default: "11000"

Value

A list with four slots: tuning data frame with the results of the tuning analysis; tuning.long, a long version of the previous data frame; tuning.plot, ggplot of tuning.long; ranger.arguments, a list ready to be used as the argument ranger.arguments in rf_repeat() or rf_spatial().

Details

The tuning method "oob" uses as reference the RMSE computed on the out-of-bag data, while the method "spatial.cv" uses RMSE computed on spatially independent data unseen by the model. The RMSE values of the latter method will always be higher, but inform better about the capacity of the combinations of hyperparameters to yield more general models.

Examples

# \donttest{
if(interactive()){

data(plant_richness_df)
data(distance_matrix)

tuning <- rf_tuning(
  data = plant_richness_df,
  dependent.variable.name = "richness_species_vascular",
  predictor.variable.names = colnames(plant_richness_df)[5:21],
  method = "oob",
  n.cores = 1,
  verbose = FALSE
)

}
# }