Applies thinning to pairs of coordinates

Resamples a set of points with x and y coordinates to impose a minimum distance among nearby points.

Usage

thinning(xy, minimum.distance = NULL)

Arguments

xy

A data frame with columns named "x" and "y" representing geographic coordinates.

minimum.distance

Numeric, minimum distance to be set between nearby points, in the same units as the coordinates of xy.

Value

A data frame with the same columns as xy with points separated by the defined minimum distance.

Details

Generally used to remove redundant points that could produce pseudo-replication, and to limit sampling bias by disaggregating clusters of points.

See also

Examples

 #load example data
 data(plant_richness_df)

 #thinning to points separated by 5 degrees
 plant_richness.thin <- thinning(
   x = plant_richness_df,
   minimum.distance = 5 #points separated by at least 5 degrees
   )

 head(plant_richness.thin)
#>   ecoregion_id          x          y richness_species_vascular bias_area_km2
#> 1            1  -66.22004 -0.4379482                      4835     268444.41
#> 2            2  -92.07368 18.6461103                      4360      26658.41
#> 3            3 -123.99060 54.9272077                      1362     140062.77
#> 4            4  -75.70215  4.8396643                      7818      31915.27
#> 5            5  -61.12626 -8.9085918                     10394     716681.87
#> 6            8  -74.19557 40.8416537                      7372      73432.58
#>   bias_species_per_record climate_aridity_index_average climate_hypervolume
#> 1              0.26999949                   -1.27163039          0.02239650
#> 2              0.14566151                    0.01279434          0.52194842
#> 3              0.12185540                    0.05070049          0.03088209
#> 4              0.16253423                   -0.57904797          0.68598777
#> 5              0.24753995                   -0.51033378          0.05199463
#> 6              0.01940485                    0.02578303          0.00118601
#>   climate_velocity_lgm_average neighbors_count neighbors_percent_shared_edge
#> 1                    7.6806232              10                     100.00000
#> 2                    7.9545641              13                      56.46095
#> 3                    6.4202598               8                     100.00000
#> 4                    0.5746547               4                      35.48473
#> 5                    6.5561073              10                     100.00000
#> 6                   37.9671501               8                      47.89004
#>   human_population_density topography_elevation_average
#> 1                7.0675297                    86.807776
#> 2               42.5908861                     6.921187
#> 3                0.8711942                  1130.220919
#> 4              218.8202463                  1872.903223
#> 5                3.0460006                   166.793692
#> 6              506.5112161                   112.525613
#>   landcover_herbs_percent_average fragmentation_cohesion fragmentation_division
#> 1                        16.22373               99.81169              0.9195281
#> 2                        53.90207               92.40274              0.9999953
#> 3                        42.66630               99.68820              0.9690934
#> 4                        54.36617               99.34487              0.9649031
#> 5                        28.21572               99.89918              0.8698819
#> 6                        56.48969               99.40405              0.9848386
#>   neighbors_area human_population human_footprint_average climate_bio1_average
#> 1      2005112.5        1897238.8               0.4305056            251.05816
#> 2       720242.4        1135405.4               8.3018495            261.08870
#> 3     15661626.9         122021.9               0.5687886             11.58566
#> 4       284154.0        6983707.7              11.9806927            173.21428
#> 5      4646036.6        2183013.4               1.9740117            251.82974
#> 6      1481083.8       37194423.7              28.2622898            112.20140
#>   climate_bio15_minimum
#> 1                    13
#> 2                    25
#> 3                    17
#> 4                    16
#> 5                    31
#> 6                     5